chore: convert to TypeScript
update ESLint configuration and add initial TypeScript definitions for MParticle API
