+++ package.json
@@ -5,16 +5,16 @@
   "license": "Apache-2.0",
   "repository": "mParticle/react-native-mparticle",
   "version": "2.7.13",
-  "main": "js/index.js",
-  "types": "js/index.d.ts",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
   "scripts": {
     "build": "tsc",
     "clean": "rm -rf lib",
     "prepare": "yarn clean && yarn build",
     "dev:pack": "yarn build && yarn pack --filename react-native-mparticle-latest.tgz",
     "dev:link": "./dev-link.sh",
-    "lint": "eslint js/**/*.{js,ts,tsx} --fix",
-    "format": "prettier --write js/**/*.{js,ts,tsx}",
+    "lint": "eslint '{js,lib}/**/*.{ts,tsx}' --fix",
+    "format": "prettier --write '{js,lib}/**/*.{ts,tsx}'",
     "test": "npm run lint"
   },
   "dependencies": {},
